Fix typo: Rename ScatterHierarchy.h file
